Massive cleanup; split out main API into "runtime" from "gobject"
We want to expose a few classes for primary public usage; MainLoop,
for example.  Having it mixed in with the other far less interesting
(but still public) APIs was ugly.  So, create a new "runtime" package
which holds those classes which are most interesting.

Rework compiler to use more Type. APIs instead of hardcoded strings.

svn path=/trunk/; revision=143