Don't skip no-shlib typelibs.  Do skip class structs in functions.
There was a change to introspection to remove the shared-library
entry for a number of .gir files and .typelibs.  In that case we just open
the current process.

We need to skip class structs in functions since we don't expose
class structs right now.  Could possibly fix by mapping to Class<?>.

svn path=/trunk/; revision=170